refactor(wallet): fix circular imports of ledger messages
